+++ docs/rules/0192/absolute-links.md
@@ -1,6 +1,6 @@
 ---
 rule:
-  aip: 192
+  aep: 192
   name: [core, '0192', absolute-links]
   summary: Comments must use absolute links
 permalink: /192/absolute-links
@@ -11,7 +11,7 @@ redirect_from:
 # Absolute links
 
 This rule attempts to enforce that every descriptor in every proto file uses
-absolute links, as mandated in [AIP-192][].
+absolute links, as mandated in [AEP-192][].
 
 ## Details
 
@@ -44,17 +44,17 @@ message Book {
 ## Disabling
 
 If you need to violate this rule, use a leading comment above the descriptor
-(and revel in the irony). Remember to also include an [aip.dev/not-precedent][]
+(and revel in the irony). Remember to also include an [aep.dev/not-precedent][]
 comment explaining why.
 
 ```proto
 // (-- api-linter: core::0192::absolute-links=disabled
-//     aip.dev/not-precedent: We need to do this because reasons. --)
+//     aep.dev/not-precedent: We need to do this because reasons. --)
 // A representation of [a book](/wiki/Book).
 message Book {
   string name = 1;
 }
 ```
 
-[aip-192]: https://aip.dev/192
-[aip.dev/not-precedent]: https://aip.dev/not-precedent
+[aep-192]: https://aep.dev/192
+[aep.dev/not-precedent]: https://aep.dev/not-precedent
