+++ docs/rules/0192/has-comments.md
@@ -1,6 +1,6 @@
 ---
 rule:
-  aip: 192
+  aep: 192
   name: [core, '0192', has-comments]
   summary: Everything in a proto file must be commented.
 permalink: /192/has-comments
@@ -11,7 +11,7 @@ redirect_from:
 # Omnipresent comments
 
 This rule enforces that every descriptor in every proto file has a _public_
-leading comment, as mandated in [AIP-192][].
+leading comment, as mandated in [AEP-192][].
 
 ## Details
 
@@ -45,17 +45,17 @@ message Book {
 ## Disabling
 
 If you need to violate this rule, use a leading comment above the descriptor
-(and revel in the irony). Remember to also include an [aip.dev/not-precedent][]
+(and revel in the irony). Remember to also include an [aep.dev/not-precedent][]
 comment explaining why.
 
 ```proto
 // A representation of a book.
 message Book {
   // (-- api-linter: core::0192::has-comments=disabled
-  //     aip.dev/not-precedent: We need to do this because reasons. --)
+  //     aep.dev/not-precedent: We need to do this because reasons. --)
   string name = 1;
 }
 ```
 
-[aip-192]: https://aip.dev/192
-[aip.dev/not-precedent]: https://aip.dev/not-precedent
+[aep-192]: https://aep.dev/192
+[aep.dev/not-precedent]: https://aep.dev/not-precedent
