+++ docs/rules/0192/trademarked-names.md
@@ -1,6 +1,6 @@
 ---
 rule:
-  aip: 192
+  aep: 192
   name: [core, '0192', trademarked-names]
   summary: Trademarked names should be used correctly.
 permalink: /192/trademarked-names
@@ -11,7 +11,7 @@ redirect_from:
 # Trademarked names
 
 This rule enforces trademarked names in public comments are not abbreviated and
-follow the trademark owner's branding style, as mandated in [AIP-192][].
+follow the trademark owner's branding style, as mandated in [AEP-192][].
 
 ## Details
 
@@ -70,19 +70,19 @@ message Book {
 ## Disabling
 
 If you need to violate this rule, use a leading comment above the descriptor.
-Remember to also include an [aip.dev/not-precedent][] comment explaining why.
+Remember to also include an [aep.dev/not-precedent][] comment explaining why.
 
 ```proto
 message Book {
   string name = 1;
 
   // (-- api-linter: core::0192::trademarked-names=disabled
-  //     aip.dev/not-precedent: We need to do this because reasons. --)
+  //     aep.dev/not-precedent: We need to do this because reasons. --)
   // A repository containing Markdown files for each chapter of
   // the book on Github.
   string github_repo = 2;
 }
 ```
 
-[aip-192]: https://aip.dev/192
-[aip.dev/not-precedent]: https://aip.dev/not-precedent
+[aep-192]: https://aep.dev/192
+[aep.dev/not-precedent]: https://aep.dev/not-precedent
