+++ rules/aep0192/has_comments.go
@@ -1,4 +1,4 @@
-// Copyright 2021 Google LLC
+// Copyright 2019 Google LLC
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,33 +12,29 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package aip0192
+package aep0192
 
 import (
-	"regexp"
-	"strings"
+	"fmt"
 
 	"github.com/googleapis/api-linter/lint"
+	"github.com/googleapis/api-linter/locations"
 	"github.com/googleapis/api-linter/rules/internal/utils"
 	"github.com/jhump/protoreflect/desc"
 )
 
-var noMarkdownTables = &lint.DescriptorRule{
-	Name: lint.NewRuleName(192, "no-markdown-tables"),
-	LintDescriptor: func(d desc.Descriptor) []lint.Problem {
-		for _, cmt := range utils.SeparateInternalComments(d.GetSourceInfo().GetLeadingComments()).External {
-			for _, line := range strings.Split(cmt, "\n") {
-				line = strings.TrimSpace(line)
-				if table.FindString(line) != "" {
-					return []lint.Problem{{
-						Message:    "Comments should not include Markdown tables.",
-						Descriptor: d,
-					}}
-				}
-			}
+// hasComments complains if there is no comment above something.
+var hasComments = &lint.DescriptorRule{
+	Name: lint.NewRuleName(192, "has-comments"),
+	LintDescriptor: func(d desc.Descriptor) (problems []lint.Problem) {
+		comment := utils.SeparateInternalComments(d.GetSourceInfo().GetLeadingComments())
+		if len(comment.External) == 0 {
+			problems = append(problems, lint.Problem{
+				Message:    fmt.Sprintf("Missing comment over %q.", d.GetName()),
+				Descriptor: d,
+				Location:   locations.DescriptorName(d),
+			})
 		}
-		return nil
+		return
 	},
 }
-
-var table = regexp.MustCompile(`^([|] )?[-]{3,}[ ]?([|][ ]?[-]{3,}[ ]?)+[|]?$`)
