+++ rules/aep0192/no_markdown_tables_test.go
@@ -12,38 +12,37 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package aip0192
+package aep0192
 
 import (
-	"strings"
 	"testing"
 
 	"github.com/googleapis/api-linter/rules/internal/testutils"
 )
 
-func TestNoMarkdownHeadings(t *testing.T) {
-	problem := testutils.Problems{{Message: "Markdown headings"}}
+func TestNoMarkdownTables(t *testing.T) {
+	problem := testutils.Problems{{Message: "Markdown tables"}}
 	for _, test := range []struct {
 		name     string
-		comment  string
+		Comment  string
 		problems testutils.Problems
 	}{
 		{"Valid", "foo bar baz", nil},
-		{"ValidMidline", "foo bar # baz", nil},
-		{"H1", "# foo bar baz", problem},
-		{"H2", "## foo bar baz", problem},
-		{"H3", "### foo bar baz", problem},
-		{"H4", "#### foo bar baz", problem},
-		{"InvalidThirdLine", "foo bar baz\n\n# spam eggs", problem},
+		{"ValidMidline", "foo - bar", nil},
+		{"ValidPipe", "foo | bar", nil},
+		{"WeirdButLegal", "|-|-|", nil},
+		{"TableSeparator", "--- | ---", problem},
+		{"BoundedTableSeparator", "| --- | --- |", problem},
+		{"MoreColumnTable", "--- | --- | --- | --- | ---", problem},
+		{"BiggerColumnTable", "------------ | -----", problem},
 	} {
 		t.Run(test.name, func(t *testing.T) {
-			cmt := "// " + strings.ReplaceAll(test.comment, "\n", "\n// ")
 			f := testutils.ParseProto3Tmpl(t, `
-			  {{.Comment}}
+			  // {{.Comment}}
 			  message Foo {}
-			`, struct{ Comment string }{cmt})
+			`, test)
 			m := f.GetMessageTypes()[0]
-			if diff := test.problems.SetDescriptor(m).Diff(noMarkdownHeadings.Lint(f)); diff != "" {
+			if diff := test.problems.SetDescriptor(m).Diff(noMarkdownTables.Lint(f)); diff != "" {
 				t.Errorf(diff)
 			}
 		})
